requirements.txt
paddlepaddle
pytorch-lightning==0.6.0
PyYAML==5.1.2
tensorboard==2.1.0
tensorboardX==1.6
terminado==0.8.1
test-tube==0.7.0
torch==1.2.0
torchfile==0.1.0
torchnet==0.0.4
torchsummary==1.5.1
torchvision==0.4.0

numpy~=1.21.3
Pillow~=8.4.0
imageio~=2.9.0
pathlib~=1.0.1